NER system works
In my job-hunting pipeline, I now have it so that it can scrape an arbitrary job posting and extract the basic requirements. I had to hand label hundreds of navigable html strings, then train an LDA, Logistic Regression, and a NER system on them, but it works, and works fast.